Convert to functional component
In Vue 2 [functional components](https://vuejs.org/v2/guide/render-function.html#Functional-Components) has a much smaller memory footprint. As this component  already uses a render function and does not use any feature missing from a functional component, I think it is worth to change.

Note that I removed the data property, but it wasn't being used through `this` in any part of the file. As it uses a render function it can reference the JS object directly without a need to make the component stateful.